backend/django_backend/requirements.txt
Django==3.2.5
psycopg2-binary==2.9.1
djangorestframework==3.12.4
PyJWT==2.3.0
django-cors-headers==3.7.0
channels==4.0.0
uvicorn[standard]==0.15.0
requests==2.28.0